+++ src/Framework/Config/ConfigException.php
@@ -4,17 +4,18 @@ declare(strict_types=1);
 
 namespace Gacela\Framework\Config;
 
+use Gacela\Framework\Config\GacelaFileConfig\GacelaFileConfigItem;
 use RuntimeException;
 
 final class ConfigException extends RuntimeException
 {
     /**
-     * @param array<string,GacelaJsonConfigItem> $configs
+     * @param array<string,GacelaFileConfigItem> $configs
      */
     public static function notFound(string $type, array $configs): self
     {
         return new self(sprintf(
-            'Config type not found in gacela.json: "%s". Only found: %s',
+            'Config-type not found in the Gacela config file: "%s". Only found: %s',
             $type,
             implode(', ', array_keys($configs))
         ));