+++ src/Framework/Config/GacelaFileConfig/GacelaFileConfigItem.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Gacela\Framework\Config\GacelaFileConfig;
+
+interface GacelaFileConfigItem
+{
+    public function path(): string;
+
+    public function pathLocal(): string;
+}