+++ src/Framework/Config/GacelaFileConfigFactoryInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Gacela\Framework\Config;
+
+use Gacela\Framework\Config\GacelaFileConfig\GacelaFileConfig;
+
+interface GacelaFileConfigFactoryInterface
+{
+    public function createGacelaFileConfig(): GacelaFileConfig;
+}