+++ tests/Integration/Framework/UsingEnvConfig/gacela.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+    'config' => [
+        'type'       => 'env',
+        'path'       => 'config/.env*',
+        'path_local' => 'config/.env.local.dist',
+    ],
+];