+++ tests/Integration/Framework/UsingPhpConfig/gacela.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+    'config' => [
+        'type'       => 'php',
+        'path'       => 'config/*.php',
+        'path_local' => 'config/local.php',
+    ],
+];