+++ tests/Unit/Framework/Config/ConfigInitTest.php
@@ -6,8 +6,8 @@ namespace GacelaTest\Unit\Framework\Config;
 
 use Gacela\Framework\Config\ConfigInit;
 use Gacela\Framework\Config\ConfigReaderInterface;
-use Gacela\Framework\Config\GacelaJsonConfig;
-use Gacela\Framework\Config\GacelaJsonConfigFactoryInterface;
+use Gacela\Framework\Config\GacelaFileConfig\GacelaJsonConfig;
+use Gacela\Framework\Config\GacelaFileConfigFactoryInterface;
 use Gacela\Framework\Config\PathFinderInterface;
 use PHPUnit\Framework\TestCase;
 
@@ -15,9 +15,9 @@ final class ConfigInitTest extends TestCase
 {
     public function test_no_config(): void
     {
-        $gacelaJsonConfigCreator = $this->createStub(GacelaJsonConfigFactoryInterface::class);
+        $gacelaJsonConfigCreator = $this->createStub(GacelaFileConfigFactoryInterface::class);
         $gacelaJsonConfigCreator
-            ->method('createGacelaJsonConfig')
+            ->method('createGacelaFileConfig')
             ->willReturn(GacelaJsonConfig::withDefaults());
 
         $readers = [
@@ -36,9 +36,9 @@ final class ConfigInitTest extends TestCase
 
     public function test_one_reader_linked_to_unsupported_type_is_ignored(): void
     {
-        $gacelaJsonConfigCreator = $this->createStub(GacelaJsonConfigFactoryInterface::class);
+        $gacelaJsonConfigCreator = $this->createStub(GacelaFileConfigFactoryInterface::class);
         $gacelaJsonConfigCreator
-            ->method('createGacelaJsonConfig')
+            ->method('createGacelaFileConfig')
             ->willReturn(GacelaJsonConfig::withDefaults());
 
         $pathFinder = $this->createMock(PathFinderInterface::class);
@@ -60,9 +60,9 @@ final class ConfigInitTest extends TestCase
 
     public function test_no_readers_returns_empty_array(): void
     {
-        $gacelaJsonConfigCreator = $this->createStub(GacelaJsonConfigFactoryInterface::class);
+        $gacelaJsonConfigCreator = $this->createStub(GacelaFileConfigFactoryInterface::class);
         $gacelaJsonConfigCreator
-            ->method('createGacelaJsonConfig')
+            ->method('createGacelaFileConfig')
             ->willReturn(GacelaJsonConfig::withDefaults());
 
         $pathFinder = $this->createMock(PathFinderInterface::class);
@@ -82,9 +82,9 @@ final class ConfigInitTest extends TestCase
 
     public function test_read_single_config(): void
     {
-        $gacelaJsonConfigCreator = $this->createStub(GacelaJsonConfigFactoryInterface::class);
+        $gacelaJsonConfigCreator = $this->createStub(GacelaFileConfigFactoryInterface::class);
         $gacelaJsonConfigCreator
-            ->method('createGacelaJsonConfig')
+            ->method('createGacelaFileConfig')
             ->willReturn(GacelaJsonConfig::fromArray([
                 'config' => [
                     [
@@ -113,9 +113,9 @@ final class ConfigInitTest extends TestCase
 
     public function test_read_multiple_config(): void
     {
-        $gacelaJsonConfigCreator = $this->createStub(GacelaJsonConfigFactoryInterface::class);
+        $gacelaJsonConfigCreator = $this->createStub(GacelaFileConfigFactoryInterface::class);
         $gacelaJsonConfigCreator
-            ->method('createGacelaJsonConfig')
+            ->method('createGacelaFileConfig')
             ->willReturn(GacelaJsonConfig::fromArray([
                 'config' => [
                     [